🚧 他人の商品をサインインしているユーザーが勝手に削除できないようにdeleteを修正

+++ src/items/items.service.ts
@@ -54,8 +54,12 @@ export class ItemsService {
     return item;
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
     // this.items = this.items.filter((item) => item.id !== id);
+    const item = await this.findById(id);
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません。');
+    }
     await this.itemRepository.delete({ id });
   }
 }
